+++ projects/_01_ingest/cqc_api/fargate/utils/postcode_matcher.py
@@ -166,15 +166,19 @@ def join_postcode_data(
     Returns:
         Tuple[pl.LazyFrame, pl.LazyFrame]: Matched and unmatched LazyFrames.
     """
+    orig_columns = locations_lf.collect_schema().names()
+
     joined_lf = locations_lf.join(
         postcode_lf,
         [ONSClean.contemporary_ons_import_date, postcode_col],
         "left",
     )
-    matched_lf = joined_lf.filter(pl.col(ONSClean.current_cssr).is_not_null())
 
-    unmatched_lf = joined_lf.filter(pl.col(ONSClean.current_cssr).is_null())
-    unmatched_lf = unmatched_lf.select(*locations_lf.columns)
+    matched_cssr = pl.col(ONSClean.current_cssr).is_not_null()
+
+    matched_lf = joined_lf.filter(matched_cssr)
+
+    unmatched_lf = joined_lf.remove(matched_cssr).select(*orig_columns)
 
     return matched_lf, unmatched_lf